Organize Masthead: subdirectories for Desktop, Mobile, shared styles, and assets
